Avoid NPE on misconfigured styles; fixes GEOS-4002
Before, the CSS demo page selected a style and attempted to load it
without checking that the file existed.  Now, the demo page checks it
and uses an error message in place of the SLD text if it cannot be
loaded.  Also, now the default style for the layer is loaded if none
is specified as a query parameter.

git-svn-id: http://svn.codehaus.org/geoserver/branches/2.0.x@14482 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: 296ba69ef4097d264436d999a6634bc1cf9d3881 [formerly b407b30b3029b7e9063073c13fbf53f6ca18c1fb]
Former-commit-id: e290e741c24053dd14e20357cd37eb5755687435