Avoid NPE on misconfigured styles; fixes GEOS-4002
Before, the CSS demo page selected a style and attempted to load it
without checking that the file existed.  Now, the demo page checks it
and uses an error message in place of the SLD text if it cannot be
loaded.  Also, now the default style for the layer is loaded if none
is specified as a query parameter.

git-svn-id: http://svn.codehaus.org/geoserver/trunk/src@14483 ef1d6e69-97e5-0310-af46-8a06194da32a

Former-commit-id: 1878450a78914e8c7ed49b9ebb09644276a671e9 [formerly 666f071278050c2fb38ace9d9527f584830bab83]
Former-commit-id: 6b5ade4d16bccd54b32a57184f1883982a2c0684